secure-api-gateway/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
cffi>=1.0
pydantic==2.7.4
python-multipart==0.0.9
